Add support for framebuffer operations
A framebuffer is defined as two textures, one for color and one for
depth, which can be used as targets for drawing/reading.

To implement the framebuffer, some bugs in opengl's texture class were
fixed + support for getting a texture's width/height was added, along
with a new api function to create a framebuffer of a specified width and
height.